backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
azure-storage-blob==12.18.3
azure-ai-documentintelligence==1.0.0
azure-identity==1.14.1
pydantic==2.4.2
python-dotenv==1.0.0
httpx==0.25.1
openai==1.3.3
pytest==7.4.3
tenacity==8.2.3
supervisor==4.2.5 